move accountresolver back into device manager file
